fix pre-commit error
