+++ docs/docs/api_ref/research_town.utils.md
@@ -18,7 +18,7 @@
 
 ### research_town.utils.agent_prompter.review_literature_prompting(profile: dict[str, str], papers: list[dict[str, str]], domains: list[str], contexts: list[str], model_name: str, prompt_template: dict[str, str | list[str]], return_num: int | None = 1, max_token_num: int | None = 512, temperature: float | None = 0.0, top_p: float | None = None, stream: bool | None = None) → list[str]
 
-### research_town.utils.agent_prompter.write_bio_prompting(publication_info: str, prompt_template: dict[str, str | list[str]], model_name: str = 'together_ai/mistralai/Mixtral-8x7B-Instruct-v0.1', return_num: int | None = 1, max_token_num: int | None = 512, temperature: float | None = 0.0, top_p: float | None = None, stream: bool | None = None) → list[str]
+### research_town.utils.profile_collector.write_bio_prompting(publication_info: str, prompt_template: dict[str, str | list[str]], model_name: str = 'together_ai/mistralai/Mixtral-8x7B-Instruct-v0.1', return_num: int | None = 1, max_token_num: int | None = 512, temperature: float | None = 0.0, top_p: float | None = None, stream: bool | None = None) → list[str]
 
 Write bio based on personal research history
