+++ research_town/dbs/db_profile.py
@@ -3,9 +3,11 @@ from typing import Any, List, Optional, TypeVar, get_type_hints
 from transformers import BertModel, BertTokenizer
 
 from ..configs import Config
-from ..utils.agent_prompter import write_bio_prompting
 from ..utils.logger import logger
-from ..utils.profile_collector import collect_publications_and_coauthors
+from ..utils.profile_collector import (
+    collect_publications_and_coauthors,
+    write_bio_prompting,
+)
 from ..utils.retriever import get_embed, rank_topk
 from .data import Data, Profile, Proposal
 from .db_base import BaseDB
