+++ research_town/utils/agent_prompter.py
@@ -18,33 +18,6 @@ from .string_mapper import (
 
 
 @beartype
-def write_bio_prompting(
-    publication_info: str,
-    prompt_template: Dict[str, Union[str, List[str]]],
-    model_name: str = 'together_ai/mistralai/Mixtral-8x7B-Instruct-v0.1',
-    return_num: Optional[int] = 1,
-    max_token_num: Optional[int] = 512,
-    temperature: Optional[float] = 0.0,
-    top_p: Optional[float] = None,
-    stream: Optional[bool] = None,
-) -> List[str]:
-    """
-    Write bio based on personal research history
-    """
-    template_input = {'publication_info': publication_info}
-    messages = openai_format_prompt_construct(prompt_template, template_input)
-    return model_prompting(
-        model_name,
-        messages,
-        return_num=return_num,
-        max_token_num=max_token_num,
-        temperature=temperature,
-        top_p=top_p,
-        stream=stream,
-    )
-
-
-@beartype
 def review_literature_prompting(
     profile: Dict[str, str],
     papers: List[Dict[str, str]],
