+++ research_town/utils/profile_collector.py
@@ -1,6 +1,10 @@
 from arxiv import Client, Search
-from beartype.typing import Any, Dict, List, Tuple
+from beartype import beartype
+from beartype.typing import Any, Dict, List, Optional, Tuple, Union
 from tqdm import tqdm
+
+from .model_prompting import model_prompting
+from .prompt_constructor import openai_format_prompt_construct
 
 
 def coauthor_frequency(
@@ -39,3 +43,30 @@ def collect_publications_and_coauthors(
         abstracts.append(abstract)
     co_author_names = coauthor_filter(co_authors, limit=10)
     return abstracts, co_author_names
+
+
+@beartype
+def write_bio_prompting(
+    publication_info: str,
+    prompt_template: Dict[str, Union[str, List[str]]],
+    model_name: str = 'together_ai/mistralai/Mixtral-8x7B-Instruct-v0.1',
+    return_num: Optional[int] = 1,
+    max_token_num: Optional[int] = 512,
+    temperature: Optional[float] = 0.0,
+    top_p: Optional[float] = None,
+    stream: Optional[bool] = None,
+) -> List[str]:
+    """
+    Write bio based on personal research history
+    """
+    template_input = {'publication_info': publication_info}
+    messages = openai_format_prompt_construct(prompt_template, template_input)
+    return model_prompting(
+        model_name,
+        messages,
+        return_num=return_num,
+        max_token_num=max_token_num,
+        temperature=temperature,
+        top_p=top_p,
+        stream=stream,
+    )
